Moved stuff to common, cleaned up pio config
