+++ TARS/platformio.ini
@@ -5,25 +5,10 @@
 ; Authors:
 ;   Ayberk Yaraneri
 ;   Gautam Dayal
+;   Magilan Sendhil
 ;
 ; PlatformIO Documentation:
 ; https://docs.platformio.org/page/projectconf.html
-
-
-; #############################################################################
-; Common settings for all environments
-[common]
-src_dir = ${workspacedir}
-default_src_filter = 
-  +<*> 
-  -<.git/> 
-  -<.svn/> 
-  -<example/> 
-  -<examples/> 
-  -<test/> 
-  -<tests/> 
-  -<test_target/>
-  -<test_local/>
 
 
 ; #############################################################################
@@ -34,22 +19,15 @@ board = teensy41
 framework = arduino
 upload_protocol = teensy-cli
 extra_scripts = extra_script.py
-; turn off all the warnings from eigen
-build_flags = -Wno-deprecated-declarations -Wno-unused-local-typedefs -Wno-reorder
+build_flags = -Wno-deprecated-declarations -Wno-unused-local-typedefs -Wno-reorder ; turn off all the warnings from eigen
 lib_deps = 
   PaulStoffregen/PWMServo
   greiman/ChRt@^1.3.0
   sparkfun/SparkFun LSM9DS1 IMU@^2.0.0
 
-build_src_filter = -<mcu_hybrid/> +<mcu_main/> 
+build_src_filter =  +<mcu_main/> +<common>
 test_ignore = test_local 
 ; lib_ldf_mode = chain+
-
-;uncomment this to compile on jetson nano:
-;platform_packages =
-;   toolchain-gccarmnoneeabi@https://bintray.com/platformio/tool-packages/download_file?file_path=5966cc5-toolchain-gccarmnoneeabi-linux_aarch64-1.90301.200702.tar.gz
-;   tool-teensy@https://github.com/maxgerhardt/pio-tool-teensy-arm/archive/master.zip
-
 
 ; #############################################################################
 ; Telemetry MCU Build Environment 
@@ -58,7 +36,3 @@ test_ignore = test_local
 ; #############################################################################
 ; Power Management MCU Build Environment 
 
-;uncomment this to compile on jetson nano:
-;platform_packages =
-;   toolchain-gccarmnoneeabi@https://bintray.com/platformio/tool-packages/download_file?file_path=5966cc5-toolchain-gccarmnoneeabi-linux_aarch64-1.90301.200702.tar.gz
-;   tool-teensy@https://github.com/maxgerhardt/pio-tool-teensy-arm/archive/master.zip
