+++ TARS/src/mcu_main/finite-state-machines/RocketFSMBase.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "FifoBuffer.h"
+#include "common/FifoBuffer.h"
 #include "common/packet.h"
 
 class RocketFSMBase {
