+++ TARS/src/mcu_main/telemetry.h
@@ -5,7 +5,7 @@
 
 #include <array>
 
-#include "MessageQueue.h"
+#include "common/MessageQueue.h"
 #include "common/packet.h"
 #include "mcu_main/pins.h"
